requirements.txt
frida-tools>=12.0.0
mitmproxy>=10.0.0
requests>=2.31.0
pyyaml>=6.0.1
colorama>=0.4.6
androguard>=4.0.0
rich>=13.7.0
tabulate>=0.9.0
